BP-2181 Fix error related to value type on rounding

+++ src/Helpers/CheckoutHelper.php
@@ -693,7 +693,7 @@ class CheckoutHelper
         // Get currency code
         $currency     = $order->getCurrency();
         $currencyCode = $currency !== null ? $currency->getIsoCode() : 'EUR';
-        $buckarooFee = round(str_replace(',','.',(float) $customFields['buckarooFee']), 2);
+        $buckarooFee = round((float)str_replace(',','.',(float) $customFields['buckarooFee']), 2);
 
         // Build the order line array
         $line = [
@@ -1970,7 +1970,7 @@ class CheckoutHelper
     public function getBuckarooFee($buckarooKey, string $salesChannelId = null)
     {
         if($buckarooFee = $this->getSetting($buckarooKey, $salesChannelId)){
-            return round(str_replace(',','.',$buckarooFee), 2);
+            return round((float)str_replace(',','.',$buckarooFee), 2);
         }
         return false;
     }
